+++ src/renderer/shared/enums/fontFamily.ts
@@ -1,0 +1,7 @@
+export enum FontFamily {
+  helvetica = 'Helvetica',
+  timesRoman = 'Times-Roman',
+  courier = 'Courier',
+  roboto = 'Roboto',
+  inter = 'Inter'
+}