+++ History.md
@@ -6,6 +6,7 @@ New features & improvements
 
 - Automatically prefill issue date
 - Automatically prefill invoice numbers
+- Font family selection added. Supported fonts: Helvetica, Times-Roman, Courier, Roboto, Inter
 
 ## 2026-02-13, version 2.0.2